Fix pytest-inmanta-extensions
